requirements.txt
bandit==1.5.1
Faker==1.0.2
flake8==3.7.7
Flask==1.0.2
flask-migrate==2.4.0
flask-request-id-header==0.1.1
flask-reverse-proxy-fix==0.2.1
flask-sqlalchemy==2.3.2
marshmallow-jsonapi==0.21.0
psycopg2-binary==2.7.7
python-dotenv==0.10.1
sentry-sdk[flask]==0.7.6
str2bool==1.1
waitress==1.2.0
ulid-py==0.0.9
